msm: kgsl: Wake GPU upon receiving an ioctl rather than upon touch input
Waking the GPU upon touch wastes power when the screen is being touched
in a way that does not induce animation or any actual need for GPU usage.
Instead of preemptively waking the GPU on touch input, wake it up upon
receiving a IOCTL_KGSL_GPU_COMMAND ioctl since it is a sign that the GPU
will soon be needed.

[Kazuki: Port to v5.4]
Change-Id: I848ac0b00f72f7bb2b08b478e86968ea8c595f3c

Change-Id: Ie012addbd39da4be1f87e8009356f503d9c554ff